Move `steps` out of `pipeline` (#384)
* Move `RuntimeParameter`

* Move `steps`
