+++ tests/integration/test_pipe_llms.py
@@ -18,10 +18,10 @@ from typing import Any, Dict, Generator, List
 from distilabel.llm.huggingface.transformers import TransformersLLM
 from distilabel.llm.openai import OpenAILLM
 from distilabel.pipeline.local import Pipeline
-from distilabel.pipeline.step.base import Step
-from distilabel.pipeline.step.generators.huggingface import LoadHubDataset
-from distilabel.pipeline.step.task.generation import TextGeneration
-from distilabel.pipeline.step.typing import RuntimeParameter, StepInput
+from distilabel.steps.base import RuntimeParameter, Step
+from distilabel.steps.generators.huggingface import LoadHubDataset
+from distilabel.steps.task.generation import TextGeneration
+from distilabel.steps.typing import StepInput
 
 
 class RenameColumns(Step):
